Add files via upload
Above picture represents a question that was asked in one of the coding competitions that was held by google.
Problem: So we have to develop a program that would generate right condition to win the case by choosing rock,paper and scissor. At the answer should be like that all last we should win from all three other users according to the points that we score in every round.
So, this the code that i developed to solve this problem. So we can solve this problem in many ways or using other logics . I also came up with two logics but i use this one because it toke lesser lines of code to achieve it.
I have written the code in c language.